requirements.txt
langchain
langchain-groq
pandas
streamlit
python-dotenv
PyJWT
altair
faiss-cpu
langchain-community
langchain-huggingface
sentence-transformers